third_party/blink/web_tests/external/wpt/fetch/api/basic/scheme-blob.sub.any.worker-expected.txt

This is a testharness.js-based test.
PASS Fetching [GET] URL.createObjectURL(blob) is OK
PASS Fetching [GET] blob:http://www.web-platform.test:8001/ is KO
PASS Fetching [POST] URL.createObjectURL(blob) is KO
PASS Fetching [OPTIONS] URL.createObjectURL(blob) is KO
PASS Fetching [HEAD] URL.createObjectURL(blob) is KO
PASS Fetching [PUT] URL.createObjectURL(blob) is KO
PASS Fetching [DELETE] URL.createObjectURL(blob) is KO
PASS Fetching [INVALID] URL.createObjectURL(blob) is KO
PASS Fetching [GET] blob:not-backed-by-a-blob/ is KO
FAIL Fetching URL.createObjectURL(empty_blob) is OK assert_equals: Content-Type is null expected (string) "" but got (object) null
FAIL Fetching URL.createObjectURL(empty_type_blob) is OK assert_equals: Content-Type is null expected (string) "" but got (object) null
PASS Fetching URL.createObjectURL(empty_data_blob) is OK
FAIL Blob content is not sniffed for a content type [image/png] assert_equals: Content-Type is null expected (string) "" but got (object) null
FAIL Blob content is not sniffed for a content type [text/xml] assert_equals: Content-Type is null expected (string) "" but got (object) null
FAIL Set content type to the empty string for slice with invalid content type assert_equals: expected (string) "" but got (object) null
FAIL Set content type to the empty string for slice with no content type  assert_equals: expected (string) "" but got (object) null
FAIL Blob.slice should not sniff the content for a content type assert_equals: expected (string) "" but got (object) null
Harness: the test ran to completion.
